test: stabilize creator release readiness
- Split TasksView 99-test single file into 5 isolated files (page/creator-content/release/performance/release-schedule, 99 tests total) sharing tasks-view-test-helpers.ts; each file uses jsdom env with afterEach cleanup (cleanup + restore fetch/confirm + restoreAllMocks + useRealTimers) to stop cross-test memory accumulation.
- Fix creator content selector multi-match with findAllByText/getAllByText.
- packaged: extract platform-agnostic isPackagedDataDirAbsolute(value, platform=process.platform) so tests assert behavior explicitly per platform instead of mocking process.platform (which is bound at load on Windows).
- packaged: make desktop-project-root-gate symlink tests capability-aware (skip when directory symlink privilege is unavailable, e.g. Windows without developer mode); non-symlink assertions always run.
- docs: add CW-07B acceptance report.

Red lines honored: no app src/daemon/contracts/dep/lockfile changes, no real API/.od/assets, no merge/push.
