+++ apps/web/tests/components/TasksView.page.test.tsx
@@ -2,6 +2,7 @@
 
 import { cleanup, fireEvent, render, screen, waitFor, within } from '@testing-library/react';
 import { afterEach, describe, expect, it, vi } from 'vitest';
+import { mockTasksViewFetch, countWriteRequests } from './tasks-view-test-helpers';
 import { readFileSync } from 'node:fs';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
@@ -47,208 +48,6 @@ type CreatorProjectData = {
 const originalFetch = globalThis.fetch;
 const originalConfirm = window.confirm;
 
-function mockTasksViewFetch({
-  routines = [],
-  creatorProjects = [],
-  creatorRuns = [],
-  creatorProjectData = {},
-  creatorMediaData = {},
-  creatorContentData = {},
-  creatorMediaFailures = [],
-  creatorContentFailures = [],
-  creatorReleaseData = {},
-  creatorReleaseFailures = [],
-  creatorPerformanceData = {},
-  creatorPerformanceFailures = [],
-  creatorPerformanceCreateError = false,
-}: {
-  routines?: Routine[];
-  creatorProjects?: Project[];
-  creatorRuns?: ChatRunStatusResponse[];
-  creatorProjectData?: Record<string, CreatorProjectData>;
-  creatorMediaData?: Record<string, { assets: Array<{ id: string; fileName: string; kind: string; relativePath: string; availability: string }>; taskLinks: Array<{ taskId: string; assetId: string }> }>;
-  creatorContentData?: Record<string, CreatorContentProjectData>;
-  creatorMediaFailures?: string[];
-  creatorContentFailures?: string[];
-  creatorReleaseData?: Record<string, CreatorReleasePackageData>;
-  creatorReleaseFailures?: string[];
-  creatorPerformanceData?: Record<string, CreatorPerformanceProjectData>;
-  creatorPerformanceFailures?: string[];
-  creatorPerformanceCreateError?: boolean;
-} = {}) {
-  const releaseStore: Record<string, CreatorReleasePackage[]> = {};
-  for (const [projectId, data] of Object.entries(creatorReleaseData)) {
-    releaseStore[projectId] = Array.isArray(data.releasePackages) ? data.releasePackages.map((release) => ({ ...release })) : [];
-  }
-  const performanceStore: Record<string, CreatorPerformanceSnapshot[]> = {};
-  for (const [projectId, data] of Object.entries(creatorPerformanceData)) {
-    performanceStore[projectId] = Array.isArray(data.snapshots) ? data.snapshots.map((snapshot) => ({ ...snapshot })) : [];
-  }
-  globalThis.fetch = vi.fn(async (input: RequestInfo | URL, init?: RequestInit) => {
-    const url = input.toString();
-    if (url === '/api/routines' && (!init || init.method === undefined)) {
-      return new Response(JSON.stringify({ routines }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    if (url === '/api/projects' && (!init || init.method === undefined)) {
-      return new Response(JSON.stringify({ projects: creatorProjects }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    if (url === '/api/runs' && (!init || init.method === undefined)) {
-      return new Response(JSON.stringify({ runs: creatorRuns }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    const creatorRead = /^\/api\/projects\/([^/]+)\/creator-workbench$/.exec(url);
-    if (creatorRead && (!init || init.method === undefined)) {
-      const projectId = decodeURIComponent(creatorRead[1]!);
-      return new Response(JSON.stringify(creatorProjectData[projectId] ?? { tasks: [], activities: [] }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    const creatorMediaRead = /^\/api\/projects\/([^/]+)\/creator-media-assets$/.exec(url);
-    if (creatorMediaRead && (!init || init.method === undefined)) {
-      const projectId = decodeURIComponent(creatorMediaRead[1]!);
-      if (creatorMediaFailures.includes(projectId)) return new Response(JSON.stringify({ error: 'media unavailable' }), { status: 503 });
-      return new Response(JSON.stringify(creatorMediaData[projectId] ?? { roots: [], assets: [], taskLinks: [] }), { status: 200, headers: { 'content-type': 'application/json' } });
-    }
-    const creatorContentRead = /^\/api\/projects\/([^/]+)\/creator-content$/.exec(url);
-    if (creatorContentRead && (!init || init.method === undefined)) {
-      const projectId = decodeURIComponent(creatorContentRead[1]!);
-      if (creatorContentFailures.includes(projectId)) return new Response(JSON.stringify({ error: 'content unavailable' }), { status: 503 });
-      return new Response(JSON.stringify(creatorContentData[projectId] ?? { contentProjects: [] }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    const releaseListRead = /^\/api\/projects\/([^/]+)\/creator-release-packages$/.exec(url);
-    if (releaseListRead && (!init || init.method === undefined)) {
-      const projectId = decodeURIComponent(releaseListRead[1]!);
-      if (creatorReleaseFailures.includes(projectId)) return new Response(JSON.stringify({ error: 'release unavailable' }), { status: 503 });
-      return new Response(JSON.stringify({ releasePackages: releaseStore[projectId] ?? [] }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    if (releaseListRead && init?.method === 'POST') {
-      const projectId = decodeURIComponent(releaseListRead[1]!);
-      const body = JSON.parse(String(init.body)) as { contentId: string; platform: string; title: string; coverAssetId?: string; exportAssetId?: string };
-      const id = `creator-release:${(releaseStore[projectId]?.length ?? 0) + 1}`;
-      const now = new Date().toISOString();
-      const pkg: CreatorReleasePackage = {
-        id, projectId, contentId: body.contentId, platform: body.platform as CreatorReleasePlatform,
-        status: 'draft', title: body.title, description: '', tags: [],
-        checklist: { contentComplete: false, exportConfirmed: false, coverConfirmed: false, metadataConfirmed: false, platformConfirmed: false },
-        createdAt: now, updatedAt: now,
-        ...(body.coverAssetId ? { coverAssetId: body.coverAssetId } : {}),
-        ...(body.exportAssetId ? { exportAssetId: body.exportAssetId } : {}),
-      };
-      releaseStore[projectId] = [...(releaseStore[projectId] ?? []), pkg];
-      return new Response(JSON.stringify({ releasePackage: pkg }), { status: 201, headers: { 'content-type': 'application/json' } });
-    }
-    const releaseExportRead = /^\/api\/projects\/([^/]+)\/creator-release-packages\/([^/]+)\/export$/.exec(url);
-    if (releaseExportRead && (!init || init.method === undefined || init.method === 'GET')) {
-      const projectId = decodeURIComponent(releaseExportRead[1]!);
-      const releaseId = decodeURIComponent(releaseExportRead[2]!);
-      const release = (releaseStore[projectId] ?? []).find((candidate) => candidate.id === releaseId);
-      if (!release) return new Response(JSON.stringify({ error: 'not found' }), { status: 404, headers: { 'content-type': 'application/json' } });
-      const exportDoc = {
-        ...release,
-        content: { id: release.contentId, title: release.contentId },
-        coverAsset: release.coverAssetId ? { id: release.coverAssetId, availability: 'available' } : null,
-        exportAsset: release.exportAssetId ? { id: release.exportAssetId, availability: 'available' } : null,
-      };
-      return new Response(JSON.stringify(exportDoc), { status: 200, headers: { 'content-type': 'application/json' } });
-    }
-    const releaseItem = /^\/api\/projects\/([^/]+)\/creator-release-packages\/([^/]+)$/.exec(url);
-    if (releaseItem && (init?.method === 'PATCH' || init?.method === 'DELETE')) {
-      const projectId = decodeURIComponent(releaseItem[1]!);
-      const releaseId = decodeURIComponent(releaseItem[2]!);
-      const store = releaseStore[projectId] ?? [];
-      const existing = store.find((candidate) => candidate.id === releaseId);
-      if (!existing) return new Response(JSON.stringify({ error: 'not found' }), { status: 404, headers: { 'content-type': 'application/json' } });
-      if (init.method === 'DELETE') {
-        releaseStore[projectId] = store.filter((candidate) => candidate.id !== releaseId);
-        return new Response(null, { status: 204 });
-      }
-      const body = JSON.parse(String(init.body)) as Record<string, unknown>;
-      const merged: CreatorReleasePackage = { ...existing, ...(body as Partial<CreatorReleasePackage>), projectId, id: releaseId };
-      // Mirror the daemon status gate: ready/published require a complete checklist;
-      // published additionally requires a valid publishedAt and publishedUrl.
-      const checklist = merged.checklist ?? existing.checklist;
-      const checklistComplete = !!checklist && (Object.values(checklist) as boolean[]).every(Boolean);
-      if (merged.status === 'ready' && !checklistComplete) {
-        return new Response(JSON.stringify({ error: 'ready requires all checklist items complete' }), { status: 400, headers: { 'content-type': 'application/json' } });
-      }
-      if (merged.status === 'published') {
-        if (!checklistComplete) {
-          return new Response(JSON.stringify({ error: 'published requires all checklist items complete' }), { status: 400, headers: { 'content-type': 'application/json' } });
-        }
-        if (!merged.publishedAt || !merged.publishedUrl) {
-          return new Response(JSON.stringify({ error: 'published requires a valid publishedAt and publishedUrl' }), { status: 400, headers: { 'content-type': 'application/json' } });
-        }
-      }
-      const updated: CreatorReleasePackage = merged;
-      releaseStore[projectId] = store.map((candidate) => candidate.id === releaseId ? updated : candidate);
-      return new Response(JSON.stringify({ releasePackage: updated }), { status: 200, headers: { 'content-type': 'application/json' } });
-    }
-    const performanceListRead = /^\/api\/projects\/([^/]+)\/creator-performance-snapshots$/.exec(url);
-    if (performanceListRead && (!init || init.method === undefined || init.method === 'GET')) {
-      const projectId = decodeURIComponent(performanceListRead[1]!);
-      if (creatorPerformanceFailures.includes(projectId)) return new Response(JSON.stringify({ error: 'performance unavailable' }), { status: 503 });
-      return new Response(JSON.stringify({ snapshots: performanceStore[projectId] ?? [] }), { status: 200, headers: { 'content-type': 'application/json' } });
-    }
-    if (performanceListRead && init?.method === 'POST') {
-      const projectId = decodeURIComponent(performanceListRead[1]!);
-      if (creatorPerformanceCreateError) return new Response(JSON.stringify({ error: 'performance create failed' }), { status: 500, headers: { 'content-type': 'application/json' } });
-      const body = JSON.parse(String(init.body)) as { releaseId: string; capturedAt?: string; metrics: Record<string, number>; note?: string };
-      const now = new Date().toISOString();
-      const snapshot: CreatorPerformanceSnapshot = {
-        id: `creator-performance:${(performanceStore[projectId]?.length ?? 0) + 1}`,
-        projectId, releaseId: body.releaseId, source: 'manual',
-        capturedAt: body.capturedAt ?? now, metrics: { ...body.metrics },
-        ...(body.note && body.note.trim() ? { note: body.note.trim() } : {}),
-        createdAt: now,
-      };
-      performanceStore[projectId] = [...(performanceStore[projectId] ?? []), snapshot];
-      return new Response(JSON.stringify({ snapshot }), { status: 201, headers: { 'content-type': 'application/json' } });
-    }
-    const performanceItem = /^\/api\/projects\/([^/]+)\/creator-performance-snapshots\/([^/]+)$/.exec(url);
-    if (performanceItem && init?.method === 'DELETE') {
-      const projectId = decodeURIComponent(performanceItem[1]!);
-      const snapshotId = decodeURIComponent(performanceItem[2]!);
-      const store = performanceStore[projectId] ?? [];
-      if (!store.some((candidate) => candidate.id === snapshotId)) return new Response(JSON.stringify({ error: 'not found' }), { status: 404, headers: { 'content-type': 'application/json' } });
-      performanceStore[projectId] = store.filter((candidate) => candidate.id !== snapshotId);
-      return new Response(null, { status: 204 });
-    }
-    if (url === '/api/automation-templates' && (!init || init.method === undefined)) {
-      return new Response(JSON.stringify({ templates: [] }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    if (url === '/api/automation-proposals?status=pending-review' && (!init || init.method === undefined)) {
-      return new Response(JSON.stringify({ proposals: [] }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    if (url === '/api/automation-source-packets?limit=3' && (!init || init.method === undefined)) {
-      return new Response(JSON.stringify({ packets: [] }), {
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-      });
-    }
-    return new Response(JSON.stringify({}), { status: 404 });
-  }) as typeof fetch;
-}
 
 function makeRelease(overrides: Partial<CreatorReleasePackage> = {}): CreatorReleasePackage {
   return {
@@ -309,13 +108,6 @@ function performanceOverviewRowByReleaseId(overview: HTMLElement, releaseId: str
   return row as HTMLElement;
 }
 
-function countWriteRequests(): number {
-  const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-  return fetchMock.mock.calls.filter((call) => {
-    const method = (call[1] as RequestInit | undefined)?.method;
-    return method === 'POST' || method === 'PATCH' || method === 'DELETE' || method === 'PUT';
-  }).length;
-}
 
 // CW-06: 读取 Release schedule 区域内所有排期项的 releaseId 顺序（跨日期组，按文档顺序）。
 function releaseScheduleItemOrder(region: HTMLElement): string[] {
@@ -333,12 +125,17 @@ function releaseScheduleItemByReleaseId(region: HTMLElement, releaseId: string):
   return li as HTMLElement;
 }
 
+
 describe('TasksView page shell', () => {
   afterEach(() => {
     cleanup();
     globalThis.fetch = originalFetch;
     window.confirm = originalConfirm;
     vi.restoreAllMocks();
+    vi.useRealTimers();
+    if (typeof globalThis !== 'undefined' && typeof (globalThis as { gc?: () => void }).gc === 'function') {
+      (globalThis as { gc?: () => void }).gc!();
+    }
   });
 
   it('renders the automations page hero and summary metrics', async () => {
@@ -1815,1453 +1612,4 @@ describe('TasksView page shell', () => {
     });
   });
 
-  it('creates a content project and saves its brief, outline, storyboard, and retrospective', async () => {
-    const creatorProject: Project = {
-      id: 'project-content-1', name: '内容项目', skillId: null, designSystemId: null,
-      createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' },
-    };
-    const writes: Array<{ url: string; body: Record<string, unknown> }> = [];
-    mockTasksViewFetch({ creatorProjects: [creatorProject] });
-    const baseFetch = globalThis.fetch;
-    globalThis.fetch = vi.fn(async (input: RequestInfo | URL, init?: RequestInit) => {
-      const url = input.toString();
-      if (url === '/api/projects/project-content-1/creator-content' && init?.method === 'POST') {
-        writes.push({ url, body: JSON.parse(String(init.body)) as Record<string, unknown> });
-        return new Response(JSON.stringify({ content: {
-          id: 'creator-content:1', projectId: creatorProject.id, title: '校园黄昏短片', status: 'idea',
-          brief: {}, outline: {}, storyboardItems: [], retrospective: {}, taskIds: [],
-          createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-        } }), { status: 201, headers: { 'content-type': 'application/json' } });
-      }
-      if (url === '/api/projects/project-content-1/creator-content/creator-content%3A1' && init?.method === 'PATCH') {
-        writes.push({ url, body: JSON.parse(String(init.body)) as Record<string, unknown> });
-        return new Response(JSON.stringify({}), { status: 200, headers: { 'content-type': 'application/json' } });
-      }
-      return baseFetch(input, init);
-    }) as typeof fetch;
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.change(screen.getByLabelText('Content title'), { target: { value: '校园黄昏短片' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Create content' }));
-    fireEvent.change(await screen.findByLabelText('Brief topic'), { target: { value: '毕业前的傍晚' } });
-    fireEvent.change(screen.getByLabelText('Outline opening'), { target: { value: '下课铃' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Add storyboard item' }));
-    fireEvent.click(screen.getByRole('button', { name: 'Add storyboard item' }));
-    fireEvent.change(screen.getAllByLabelText('Storyboard purpose')[0]!, { target: { value: '建立时间' } });
-    fireEvent.change(screen.getAllByLabelText('Storyboard purpose')[1]!, { target: { value: '跟拍离开教室' } });
-    fireEvent.change(screen.getByLabelText('Retrospective learnings'), { target: { value: '保留环境声' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Save content' }));
-
-    await waitFor(() => expect(writes).toEqual(expect.arrayContaining([
-      expect.objectContaining({ url: '/api/projects/project-content-1/creator-content', body: { title: '校园黄昏短片', status: 'idea' } }),
-      expect.objectContaining({ url: '/api/projects/project-content-1/creator-content/creator-content%3A1', body: expect.objectContaining({
-        brief: expect.objectContaining({ topic: '毕业前的傍晚' }),
-        outline: expect.objectContaining({ opening: '下课铃' }),
-        storyboardItems: expect.arrayContaining([expect.objectContaining({ purpose: '建立时间' }), expect.objectContaining({ purpose: '跟拍离开教室' })]),
-        retrospective: expect.objectContaining({ learnings: '保留环境声' }),
-      }) }),
-    ])));
-  });
-
-  it('shows missing storyboard media and never offers it as a new candidate', async () => {
-    const creatorProject: Project = { id: 'project-content-media-1', name: '内容素材项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-content-media-1': { contentProjects: [{
-        id: 'creator-content:1', projectId: creatorProject.id, title: '校园黄昏短片', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [],
-        storyboardItems: [{ id: 'creator-storyboard:1', position: 1, purpose: '开场', mediaAssetIds: ['creator-media:missing'], createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z' }],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorMediaData: { 'project-content-media-1': { assets: [
-        { id: 'creator-media:missing', fileName: 'missing.jpg', kind: 'image', relativePath: 'missing.jpg', availability: 'missing' },
-        { id: 'creator-media:available', fileName: 'available.jpg', kind: 'image', relativePath: 'available.jpg', availability: 'available' },
-      ], taskLinks: [] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit content 校园黄昏短片' }));
-    expect(screen.getAllByText('Missing').length).toBeGreaterThan(0);
-    const candidates = screen.getByLabelText('Storyboard media candidate');
-    expect(within(candidates).queryByRole('option', { name: 'missing.jpg' })).toBeNull();
-      expect(within(candidates).getByRole('option', { name: 'available.jpg' })).toBeTruthy();
-  });
-
-  it('shows a missing asset hint in the content list before opening the editor', async () => {
-    const creatorProject: Project = { id: 'project-content-missing-hint-1', name: '缺失素材提示项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-content-missing-hint-1': { contentProjects: [
-        {
-          id: 'creator-content:missing', projectId: creatorProject.id, title: '有缺失素材的内容', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [],
-          storyboardItems: [{ id: 'creator-storyboard:missing', position: 1, purpose: '开场', mediaAssetIds: ['creator-media:missing'], createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z' }],
-          createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-        },
-        {
-          id: 'creator-content:ok', projectId: creatorProject.id, title: '资料齐全的内容', status: 'idea', brief: {}, outline: {}, retrospective: {}, taskIds: [],
-          storyboardItems: [{ id: 'creator-storyboard:ok', position: 1, purpose: '开场', mediaAssetIds: ['creator-media:available'], createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z' }],
-          createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-        },
-        {
-          id: 'creator-content:empty', projectId: creatorProject.id, title: '无素材的内容', status: 'idea', brief: {}, outline: {}, retrospective: {}, taskIds: [],
-          storyboardItems: [],
-          createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-        },
-      ] } },
-      creatorMediaData: { 'project-content-missing-hint-1': { assets: [
-        { id: 'creator-media:missing', fileName: 'missing.jpg', kind: 'image', relativePath: 'missing.jpg', availability: 'missing' },
-        { id: 'creator-media:available', fileName: 'available.jpg', kind: 'image', relativePath: 'available.jpg', availability: 'available' },
-      ], taskLinks: [] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-
-    // The missing-asset content surfaces the hint before the editor opens.
-    expect(await screen.findByText('1 missing asset')).toBeTruthy();
-    // Available-only and asset-less content are listed but carry no hint.
-    expect(screen.getByText('资料齐全的内容')).toBeTruthy();
-    expect(screen.getByText('无素材的内容')).toBeTruthy();
-    // Exactly one missing-asset chip exists: only the missing-media content shows it.
-    expect(screen.getAllByText(/missing asset/i).length).toBe(1);
-  });
-
-  it('asks for confirmation before deleting a content project', async () => {
-    const creatorProject: Project = { id: 'project-content-delete-1', name: '删除内容项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const deleteCalls: string[] = [];
-    const confirmSpy = vi.spyOn(window, 'confirm').mockReturnValueOnce(false).mockReturnValueOnce(true);
-    mockTasksViewFetch({ creatorProjects: [creatorProject], creatorContentData: { 'project-content-delete-1': { contentProjects: [{
-      id: 'creator-content:1', projectId: creatorProject.id, title: '校园黄昏短片', status: 'idea', brief: {}, outline: {}, storyboardItems: [], retrospective: {}, taskIds: [],
-      createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-    }] } } });
-    const baseFetch = globalThis.fetch;
-    globalThis.fetch = vi.fn(async (input: RequestInfo | URL, init?: RequestInit) => {
-      if (input.toString() === '/api/projects/project-content-delete-1/creator-content/creator-content%3A1' && init?.method === 'DELETE') {
-        deleteCalls.push(input.toString());
-        return new Response(null, { status: 204 });
-      }
-      return baseFetch(input, init);
-    }) as typeof fetch;
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Delete content 校园黄昏短片' }));
-    expect(deleteCalls).toEqual([]);
-    fireEvent.click(screen.getByRole('button', { name: 'Delete content 校园黄昏短片' }));
-    await waitFor(() => expect(deleteCalls).toEqual(['/api/projects/project-content-delete-1/creator-content/creator-content%3A1']));
-    expect(confirmSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it('degrades creator content when the content API fails but keeps tasks and media usable', async () => {
-    const creatorProject: Project = {
-      id: 'project-content-fail-1', name: '内容失败项目', skillId: null, designSystemId: null,
-      createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' },
-    };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentFailures: ['project-content-fail-1'],
-      creatorProjectData: { 'project-content-fail-1': { tasks: [{
-        id: 'creator-task:1', projectId: creatorProject.id, title: 'surviving task', stage: 'topic', status: 'todo',
-        priority: 'medium', createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }], activities: [] } },
-      creatorMediaData: { 'project-content-fail-1': { assets: [
-        { id: 'creator-media:1', fileName: 'clip.mp4', kind: 'video', relativePath: 'media/clip.mp4', availability: 'available' },
-      ], taskLinks: [] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    expect(await screen.findByText('Content unavailable for this project.')).toBeTruthy();
-    expect(screen.getAllByText('surviving task').length).toBeGreaterThan(0);
-    expect(screen.getAllByText('clip.mp4').length).toBeGreaterThan(0);
-  });
-
-  it('reads release list per project and sends a create body limited to contentId, platform, and title', async () => {
-    const creatorProject: Project = { id: 'project-release-1', name: '发布项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-release-1': { contentProjects: [{
-        id: 'creator-content:1', projectId: 'project-release-1', title: '校园短片', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const contentSelect = await screen.findByLabelText('Release content');
-    await screen.findByRole('option', { name: '校园短片' });
-    fireEvent.change(contentSelect, { target: { value: 'creator-content:1' } });
-    fireEvent.change(screen.getByLabelText('Release title'), { target: { value: '首发短片' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Create release' }));
-
-    const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-    await waitFor(() => {
-      const postCall = fetchMock.mock.calls.find(
-        (call) => String(call[0]).endsWith('/creator-release-packages') && (call[1] as RequestInit | undefined)?.method === 'POST',
-      );
-      expect(postCall).toBeTruthy();
-      const sent = JSON.parse(String((postCall![1] as RequestInit).body));
-      expect(sent).toEqual({ contentId: 'creator-content:1', platform: 'bilibili', title: '首发短片' });
-    });
-    // The new package appears in the project's release list.
-    expect(await screen.findByText('首发短片')).toBeTruthy();
-  });
-
-  it('sends null to clear optional release fields on save', async () => {
-    const creatorProject: Project = { id: 'project-release-2', name: '空字段发布', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({
-      id: 'creator-release:2', projectId: 'project-release-2', title: '待清空',
-      coverAssetId: 'creator-media:cover',
-      scheduledAt: '2025-06-01T00:00:00.000Z',
-      publishedAt: '2025-06-02T00:00:00.000Z',
-      publishedUrl: 'https://example.com/x',
-    });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-release-2': { contentProjects: [{
-        id: 'creator-content:2', projectId: 'project-release-2', title: '内容B', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorReleaseData: { 'project-release-2': { releasePackages: [release] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 待清空' }));
-    fireEvent.change(screen.getByLabelText('Edit release cover asset'), { target: { value: '' } });
-    fireEvent.change(screen.getByLabelText('Edit release scheduled at'), { target: { value: '' } });
-    fireEvent.change(screen.getByLabelText('Edit release published at'), { target: { value: '' } });
-    fireEvent.change(screen.getByLabelText('Edit release published url'), { target: { value: '' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Save release' }));
-
-    const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-    await waitFor(() => {
-      const patchCall = fetchMock.mock.calls.find(
-        (call) => String(call[0]).includes('/creator-release-packages/creator-release%3A2') && (call[1] as RequestInit | undefined)?.method === 'PATCH',
-      );
-      expect(patchCall).toBeTruthy();
-      const sent = JSON.parse(String((patchCall![1] as RequestInit).body));
-      expect(sent.coverAssetId).toBeNull();
-      expect(sent.exportAssetId).toBeNull();
-      expect(sent.scheduledAt).toBeNull();
-      expect(sent.publishedAt).toBeNull();
-      expect(sent.publishedUrl).toBeNull();
-    });
-  });
-
-  it('rejects ready/published release saves when the checklist is incomplete and preserves the editor input', async () => {
-    const creatorProject: Project = { id: 'project-release-3', name: '门禁发布', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({ id: 'creator-release:3', projectId: 'project-release-3', title: '待发布' });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-release-3': { contentProjects: [{
-        id: 'creator-content:3', projectId: 'project-release-3', title: '内容C', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorReleaseData: { 'project-release-3': { releasePackages: [release] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 待发布' }));
-    fireEvent.change(screen.getByLabelText('Edit release status'), { target: { value: 'ready' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Save release' }));
-
-    const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-    await waitFor(async () => {
-      const idx = fetchMock.mock.calls.findIndex(
-        (call) => String(call[0]).includes('/creator-release-packages/creator-release%3A3') && (call[1] as RequestInit | undefined)?.method === 'PATCH',
-      );
-      expect(idx).toBeGreaterThan(-1);
-      const res = await (fetchMock.mock.results[idx] as { value: Promise<Response> }).value;
-      expect(res.status).toBe(400);
-    });
-    // Editor input is preserved (status stays "ready", title unchanged) and an error is shown.
-    expect((screen.getByLabelText('Edit release status') as HTMLSelectElement).value).toBe('ready');
-    expect((screen.getByLabelText('Edit release title') as HTMLInputElement).value).toBe('待发布');
-    expect(screen.getByRole('alert')).toBeTruthy();
-  });
-
-  it('keeps a missing asset reference on the release and offers it as a retained candidate', async () => {
-    const creatorProject: Project = { id: 'project-release-4', name: '缺失素材发布', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({ id: 'creator-release:4', projectId: 'project-release-4', title: '带缺失素材', coverAssetId: 'creator-media:missing' });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-release-4': { contentProjects: [{
-        id: 'creator-content:4', projectId: 'project-release-4', title: '内容D', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorMediaData: { 'project-release-4': { assets: [
-        { id: 'creator-media:missing', fileName: 'missing.jpg', kind: 'image', relativePath: 'missing.jpg', availability: 'missing' },
-        { id: 'creator-media:available', fileName: 'available.jpg', kind: 'image', relativePath: 'available.jpg', availability: 'available' },
-      ], taskLinks: [] } },
-      creatorReleaseData: { 'project-release-4': { releasePackages: [release] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    // The list surfaces the missing-asset chip before opening the editor.
-    expect(await screen.findByText('1 missing asset')).toBeTruthy();
-    fireEvent.click(screen.getByRole('button', { name: 'Edit release 带缺失素材' }));
-    // The missing asset is retained as a candidate and flagged Missing.
-    const coverSelect = screen.getByLabelText('Edit release cover asset') as HTMLSelectElement;
-    expect(within(coverSelect).getByRole('option', { name: 'missing.jpg (Missing)' })).toBeTruthy();
-    const editor = document.querySelector('.creator-release-editor') as HTMLElement;
-    expect(within(editor).getByText('Missing')).toBeTruthy();
-    // Saving without changing the cover keeps the missing reference (not nulled).
-    fireEvent.click(screen.getByRole('button', { name: 'Save release' }));
-    const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-    await waitFor(() => {
-      const patchCall = fetchMock.mock.calls.find(
-        (call) => String(call[0]).includes('/creator-release-packages/creator-release%3A4') && (call[1] as RequestInit | undefined)?.method === 'PATCH',
-      );
-      expect(patchCall).toBeTruthy();
-      const sent = JSON.parse(String((patchCall![1] as RequestInit).body));
-      expect(sent.coverAssetId).toBe('creator-media:missing');
-    });
-  });
-
-  it('shows the release unavailable state when the release API fails but keeps other panels usable', async () => {
-    const creatorProject: Project = {
-      id: 'project-release-5', name: '发布失败项目', skillId: null, designSystemId: null,
-      createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' },
-    };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseFailures: ['project-release-5'],
-      creatorProjectData: { 'project-release-5': { tasks: [{
-        id: 'creator-task:release-fail', projectId: 'project-release-5', title: 'surviving task', stage: 'topic', status: 'todo',
-        priority: 'medium', createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }], activities: [] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    expect(await screen.findByText('Release unavailable for this project.')).toBeTruthy();
-    expect(screen.getAllByText('surviving task').length).toBeGreaterThan(0);
-  });
-
-  it('asks for confirmation before deleting a release and cancels on decline', async () => {
-    const creatorProject: Project = { id: 'project-release-6', name: '删除发布', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({ id: 'creator-release:6', projectId: 'project-release-6', title: '待删除' });
-    const confirmSpy = vi.spyOn(window, 'confirm').mockReturnValueOnce(false).mockReturnValueOnce(true);
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-release-6': { contentProjects: [{
-        id: 'creator-content:6', projectId: 'project-release-6', title: '内容E', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorReleaseData: { 'project-release-6': { releasePackages: [release] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Delete release 待删除' }));
-    fireEvent.click(screen.getByRole('button', { name: 'Delete release 待删除' }));
-    await waitFor(() => {
-      const deleteCalls = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls.filter(
-        (call) => String(call[0]).includes('/creator-release-packages/creator-release%3A6') && (call[1] as RequestInit | undefined)?.method === 'DELETE',
-      );
-      expect(deleteCalls.length).toBe(1);
-    });
-    expect(confirmSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it('downloads the release package as JSON and Markdown blobs', async () => {
-    const creatorProject: Project = { id: 'project-release-7', name: '导出发布', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({
-      id: 'creator-release:7', projectId: 'project-release-7', title: '导出包', contentId: 'creator-content:7',
-      description: '简介', tags: ['a', 'b'],
-      checklist: { contentComplete: true, exportConfirmed: true, coverConfirmed: false, metadataConfirmed: true, platformConfirmed: false },
-    });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorContentData: { 'project-release-7': { contentProjects: [{
-        id: 'creator-content:7', projectId: 'project-release-7', title: '内容F', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorReleaseData: { 'project-release-7': { releasePackages: [release] } },
-    });
-
-    const capturedBlobs: Blob[] = [];
-    const originalCreate = (URL as { createObjectURL?: (blob: Blob) => string }).createObjectURL;
-    Object.defineProperty(URL, 'createObjectURL', {
-      configurable: true,
-      writable: true,
-      value: (blob: Blob) => { capturedBlobs.push(blob); return 'blob:mock'; },
-    });
-
-    try {
-      render(<TasksView projects={[creatorProject]} />);
-      fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-      fireEvent.click(await screen.findByRole('button', { name: 'Edit release 导出包' }));
-      fireEvent.click(screen.getByRole('button', { name: 'Download JSON' }));
-      await waitFor(() => expect(capturedBlobs.length).toBeGreaterThanOrEqual(1));
-      const json = JSON.parse(await capturedBlobs[0]!.text()) as { id: string; title: string; content: { id: string }; checklist: CreatorReleaseChecklist };
-      expect(json.id).toBe('creator-release:7');
-      expect(json.title).toBe('导出包');
-      expect(json.content.id).toBe('creator-content:7');
-      expect(json.checklist.contentComplete).toBe(true);
-
-      fireEvent.click(screen.getByRole('button', { name: 'Download Markdown' }));
-      await waitFor(() => expect(capturedBlobs.length).toBeGreaterThanOrEqual(2));
-      const md = await capturedBlobs[1]!.text();
-      expect(md).toContain('# 导出包');
-      expect(md).toContain('- Platform: bilibili');
-      expect(md).toContain('- Tags: a, b');
-      expect(md).toContain('## Checklist');
-      expect(md).toContain('contentComplete: true');
-    } finally {
-      if (originalCreate) {
-        Object.defineProperty(URL, 'createObjectURL', { configurable: true, writable: true, value: originalCreate });
-      } else {
-        delete (URL as { createObjectURL?: unknown }).createObjectURL;
-      }
-    }
-  });
-
-  it('ships responsive release layout rules for the 960px and 640px breakpoints', () => {
-    // jsdom does not evaluate media queries, so we assert the rules exist in the stylesheet.
-    const cssPath = join(dirname(fileURLToPath(import.meta.url)), '..', '..', 'src', 'styles', 'home', 'tasks.css');
-    const css = readFileSync(cssPath, 'utf8').replace(/\r\n/g, '\n');
-    expect(css).toContain('@media (max-width: 960px) {\n  .creator-release-layout,');
-    expect(css).toContain('@media (max-width: 640px) {\n  .creator-release-create,');
-    expect(css).toContain('.creator-release-editor .creator-check {');
-  });
-
-  it('creates a performance snapshot with only the filled metric keys for a published release', async () => {
-    const creatorProject: Project = { id: 'project-perf-1', name: '表现复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:perf1', projectId: 'project-perf-1', title: '已发布包' });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-1': { releasePackages: [release] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 已发布包' }));
-    const views = await screen.findByLabelText('Performance views');
-    fireEvent.change(views, { target: { value: '100' } });
-    fireEvent.change(screen.getByLabelText('Performance likes'), { target: { value: '20' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Save snapshot' }));
-
-    const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-    await waitFor(() => {
-      const postCall = fetchMock.mock.calls.find(
-        (call) => String(call[0]).endsWith('/creator-performance-snapshots') && (call[1] as RequestInit | undefined)?.method === 'POST',
-      );
-      expect(postCall).toBeTruthy();
-      const sent = JSON.parse(String((postCall![1] as RequestInit).body));
-      expect(sent).toEqual({ releaseId: 'creator-release:perf1', metrics: { views: 100, likes: 20 } });
-    });
-    // 成功后表单清空，且新快照出现在列表中。
-    expect((screen.getByLabelText('Performance views') as HTMLInputElement).value).toBe('');
-    expect(await screen.findByText('views: 100')).toBeTruthy();
-  });
-
-  it('hides the performance snapshot form for a non-published release', async () => {
-    const creatorProject: Project = { id: 'project-perf-2', name: '草稿复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({ id: 'creator-release:perf2', projectId: 'project-perf-2', title: '草稿包', status: 'draft' });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-2': { releasePackages: [release] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 草稿包' }));
-    expect(await screen.findByText('Performance snapshots require a published release.')).toBeTruthy();
-    expect(screen.queryByLabelText('Performance views')).toBeNull();
-    expect(screen.queryByRole('button', { name: 'Save snapshot' })).toBeNull();
-  });
-
-  it('retains archived performance snapshots and permits deleting a mistaken entry', async () => {
-    const creatorProject: Project = { id: 'project-perf-archived', name: '归档复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makeRelease({ id: 'creator-release:archived', projectId: 'project-perf-archived', title: '归档包', status: 'archived' });
-    const snapshot: CreatorPerformanceSnapshot = {
-      id: 'creator-performance:archived', projectId: 'project-perf-archived', releaseId: release.id, source: 'manual',
-      capturedAt: '2026-07-04T00:00:00.000Z', metrics: { views: 8 }, createdAt: '2026-07-04T00:00:00.000Z',
-    };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-archived': { releasePackages: [release] } },
-      creatorPerformanceData: { 'project-perf-archived': { snapshots: [snapshot] } },
-    });
-    vi.spyOn(window, 'confirm').mockReturnValue(true);
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 归档包' }));
-
-    expect(await screen.findByText('Performance snapshots require a published release.')).toBeTruthy();
-    expect(screen.getByText('views: 8')).toBeTruthy();
-    expect(screen.queryByRole('button', { name: 'Save snapshot' })).toBeNull();
-    fireEvent.click(screen.getByRole('button', { name: 'Delete performance snapshot 2026-07-04T00:00:00.000Z' }));
-    await waitFor(() => expect((globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls.some(
-      (call) => String(call[0]).includes('/creator-performance-snapshots/creator-performance%3Aarchived') && (call[1] as RequestInit | undefined)?.method === 'DELETE',
-    )).toBe(true));
-  });
-
-  it('preserves performance snapshot input and shows an alert on a failed create', async () => {
-    const creatorProject: Project = { id: 'project-perf-3', name: '失败复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:perf3', projectId: 'project-perf-3', title: '已发布包3' });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-3': { releasePackages: [release] } },
-      creatorPerformanceCreateError: true,
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 已发布包3' }));
-    const views = await screen.findByLabelText('Performance views');
-    fireEvent.change(views, { target: { value: '100' } });
-    fireEvent.click(screen.getByRole('button', { name: 'Save snapshot' }));
-
-    await waitFor(() => expect(screen.getByRole('alert')).toBeTruthy());
-    expect((screen.getByLabelText('Performance views') as HTMLInputElement).value).toBe('100');
-  });
-
-  it('lists performance snapshots by capturedAt desc and shows signed deltas for shared metrics', async () => {
-    const creatorProject: Project = { id: 'project-perf-4', name: '增量复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:perf4', projectId: 'project-perf-4', title: '已发布包4' });
-    const older: CreatorPerformanceSnapshot = {
-      id: 'creator-performance:old', projectId: 'project-perf-4', releaseId: 'creator-release:perf4', source: 'manual',
-      capturedAt: '2026-07-01T00:00:00.000Z', metrics: { views: 100, likes: 10 }, createdAt: '2026-07-01T00:00:00.000Z',
-    };
-    const newer: CreatorPerformanceSnapshot = {
-      id: 'creator-performance:new', projectId: 'project-perf-4', releaseId: 'creator-release:perf4', source: 'manual',
-      capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 150, likes: 20 }, createdAt: '2026-07-03T00:00:00.000Z',
-    };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-4': { releasePackages: [release] } },
-      creatorPerformanceData: { 'project-perf-4': { snapshots: [older, newer] } },
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 已发布包4' }));
-
-    // 较新快照在前，显示相对紧邻更旧快照的增量。
-    expect(await screen.findByText('views: 150 (+50)')).toBeTruthy();
-    expect(screen.getByText('likes: 20 (+10)')).toBeTruthy();
-    // 较旧快照没有更旧的可比对象，不显示增量。
-    expect(screen.getByText('views: 100')).toBeTruthy();
-    expect(screen.getAllByText(/\(\+50\)/).length).toBe(1);
-    expect(screen.getAllByText(/\(\+10\)/).length).toBe(1);
-  });
-
-  it('confirms before deleting a performance snapshot and removes it only after confirm', async () => {
-    const creatorProject: Project = { id: 'project-perf-5', name: '删除复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:perf5', projectId: 'project-perf-5', title: '已发布包5' });
-    const snapshot: CreatorPerformanceSnapshot = {
-      id: 'creator-performance:del', projectId: 'project-perf-5', releaseId: 'creator-release:perf5', source: 'manual',
-      capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 5 }, createdAt: '2026-07-02T00:00:00.000Z',
-    };
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-5': { releasePackages: [release] } },
-      creatorPerformanceData: { 'project-perf-5': { snapshots: [snapshot] } },
-    });
-    const confirmSpy = vi.spyOn(window, 'confirm').mockReturnValueOnce(false).mockReturnValueOnce(true);
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 已发布包5' }));
-    const deleteBtn = await screen.findByRole('button', { name: 'Delete performance snapshot 2026-07-02T00:00:00.000Z' });
-    fireEvent.click(deleteBtn);
-    fireEvent.click(deleteBtn);
-
-    const fetchMock = globalThis.fetch as ReturnType<typeof vi.fn>;
-    await waitFor(() => {
-      const delCalls = fetchMock.mock.calls.filter(
-        (call) => String(call[0]).includes('/creator-performance-snapshots/creator-performance%3Adel') && (call[1] as RequestInit | undefined)?.method === 'DELETE',
-      );
-      expect(delCalls.length).toBe(1);
-    });
-    expect(confirmSpy).toHaveBeenCalledTimes(2);
-    await waitFor(() => expect(screen.queryByText('views: 5')).toBeNull());
-  });
-
-  it('degrades the performance snapshot panel for a failing project without breaking the Release panel', async () => {
-    const creatorProject: Project = { id: 'project-perf-6', name: '降级复盘', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:perf6', projectId: 'project-perf-6', title: '已发布包6' });
-    mockTasksViewFetch({
-      creatorProjects: [creatorProject],
-      creatorReleaseData: { 'project-perf-6': { releasePackages: [release] } },
-      creatorPerformanceFailures: ['project-perf-6'],
-    });
-
-    render(<TasksView projects={[creatorProject]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    fireEvent.click(await screen.findByRole('button', { name: 'Edit release 已发布包6' }));
-    expect(await screen.findByText('Performance unavailable for this project.')).toBeTruthy();
-    // Release 面板仍可用。
-    expect(screen.getByRole('button', { name: 'Save release' })).toBeTruthy();
-    expect(screen.queryByLabelText('Performance views')).toBeNull();
-  });
-
-  it('ships responsive performance snapshot layout rules for the 640px breakpoint', () => {
-    const cssPath = join(dirname(fileURLToPath(import.meta.url)), '..', '..', 'src', 'styles', 'home', 'tasks.css');
-    const css = readFileSync(cssPath, 'utf8').replace(/\r\n/g, '\n');
-    expect(css).toContain('.creator-performance {');
-    expect(css).toContain('.creator-performance-fields {');
-    expect(css).toContain('@media (max-width: 640px) {\n  .creator-release-create,\n  .creator-release-fields--top {\n    grid-template-columns: 1fr;\n  }\n  .creator-performance-fields {\n    grid-template-columns: 1fr;\n  }\n}');
-  });
-
-  it('performance overview lists only published releases across multiple projects', async () => {
-    const projectA: Project = { id: 'ov-proj-a', name: '项目甲', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const projectB: Project = { id: 'ov-proj-b', name: '项目乙', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const pubA = makePublishedRelease({ id: 'creator-release:ovA1', projectId: 'ov-proj-a', title: '甲已发布', platform: 'bilibili' });
-    const draftA = makeRelease({ id: 'creator-release:ovAd', projectId: 'ov-proj-a', title: '甲草稿', status: 'draft' });
-    const pubB = makePublishedRelease({ id: 'creator-release:ovB1', projectId: 'ov-proj-b', title: '乙已发布', platform: 'youtube' });
-    mockTasksViewFetch({
-      creatorProjects: [projectA, projectB],
-      creatorReleaseData: {
-        'ov-proj-a': { releasePackages: [pubA, draftA] },
-        'ov-proj-b': { releasePackages: [pubB] },
-      },
-      creatorPerformanceData: {
-        'ov-proj-a': { snapshots: [makeSnapshot({ id: 's-a', projectId: 'ov-proj-a', releaseId: 'creator-release:ovA1', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 10 } })] },
-        'ov-proj-b': { snapshots: [makeSnapshot({ id: 's-b', projectId: 'ov-proj-b', releaseId: 'creator-release:ovB1', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 20 } })] },
-      },
-    });
-
-    render(<TasksView projects={[projectA, projectB]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(2));
-    const order = performanceOverviewReleaseOrder(overview);
-    expect(order).toContain('creator-release:ovA1');
-    expect(order).toContain('creator-release:ovB1');
-    expect(order).not.toContain('creator-release:ovAd');
-    expect(within(overview).getByText('甲已发布')).toBeTruthy();
-    expect(within(overview).getByText('乙已发布')).toBeTruthy();
-    expect(within(overview).queryByText('甲草稿')).toBeNull();
-  });
-
-  it('performance overview selects latest and previous snapshots by capturedAt', async () => {
-    const project: Project = { id: 'ov-latest', name: '最新快照', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:ovL', projectId: 'ov-latest', title: '最新包' });
-    // 快照乱序传入：中间、最新、最旧；总览必须选出最新一条并对相邻上一条求增量。
-    const snapshots: CreatorPerformanceSnapshot[] = [
-      makeSnapshot({ id: 's-mid', projectId: 'ov-latest', releaseId: 'creator-release:ovL', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 130 } }),
-      makeSnapshot({ id: 's-new', projectId: 'ov-latest', releaseId: 'creator-release:ovL', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 150 } }),
-      makeSnapshot({ id: 's-old', projectId: 'ov-latest', releaseId: 'creator-release:ovL', capturedAt: '2026-07-01T00:00:00.000Z', metrics: { views: 100 } }),
-    ];
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-latest': { releasePackages: [release] } },
-      creatorPerformanceData: { 'ov-latest': { snapshots } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    const row = await waitFor(() => performanceOverviewRowByReleaseId(overview, 'creator-release:ovL'));
-    const tds = row.querySelectorAll('td');
-    // 最新快照时间列显示最新一条 capturedAt。
-    expect(tds[3]!.textContent).toContain('2026-07-03T00:00:00.000Z');
-    // views 相对上一条（130）为 +20，而非相对最旧的 100。
-    expect(tds[4]!.textContent?.replace(/\s+/g, '')).toBe('150+20');
-  });
-
-  it('performance overview shows positive, negative, and zero deltas', async () => {
-    const project: Project = { id: 'ov-delta', name: '增量', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:ovD', projectId: 'ov-delta', title: '增量包' });
-    const previous = makeSnapshot({ id: 's-prev', projectId: 'ov-delta', releaseId: 'creator-release:ovD', capturedAt: '2026-07-01T00:00:00.000Z', metrics: { views: 100, likes: 50, comments: 10 } });
-    const latest = makeSnapshot({ id: 's-last', projectId: 'ov-delta', releaseId: 'creator-release:ovD', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 150, likes: 30, comments: 10 } });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-delta': { releasePackages: [release] } },
-      creatorPerformanceData: { 'ov-delta': { snapshots: [previous, latest] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    const row = await waitFor(() => performanceOverviewRowByReleaseId(overview, 'creator-release:ovD'));
-    const tds = row.querySelectorAll('td');
-    expect(tds[4]!.textContent?.replace(/\s+/g, '')).toBe('150+50'); // 正增量
-    expect(tds[5]!.textContent?.replace(/\s+/g, '')).toBe('30-20'); // 负增量
-    expect(tds[6]!.textContent?.replace(/\s+/g, '')).toBe('100'); // 零增量显示 "0"
-  });
-
-  it('performance overview shows a dash for missing metrics without inferring zero and skips fake deltas', async () => {
-    const project: Project = { id: 'ov-missing', name: '缺失', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:ovM', projectId: 'ov-missing', title: '缺失包' });
-    const previous = makeSnapshot({ id: 's-mp', projectId: 'ov-missing', releaseId: 'creator-release:ovM', capturedAt: '2026-07-01T00:00:00.000Z', metrics: { views: 100, likes: 40 } });
-    // 最新快照：views 有历史可比 → 增量；comments 为有效 0 且历史缺失 → 显示 "0" 无增量；likes 缺失 → "-"。
-    const latest = makeSnapshot({ id: 's-ml', projectId: 'ov-missing', releaseId: 'creator-release:ovM', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 150, comments: 0 } });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-missing': { releasePackages: [release] } },
-      creatorPerformanceData: { 'ov-missing': { snapshots: [previous, latest] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    const row = await waitFor(() => performanceOverviewRowByReleaseId(overview, 'creator-release:ovM'));
-    const tds = row.querySelectorAll('td');
-    expect(tds[4]!.textContent?.replace(/\s+/g, '')).toBe('150+50');
-    expect(tds[5]!.textContent?.trim()).toBe('-'); // 缺失显示破折号，不补零
-    expect(tds[6]!.textContent?.trim()).toBe('0'); // 有效 0 不是缺失，且无伪造增量
-  });
-
-  it('performance overview keeps published releases without snapshots and marks them', async () => {
-    const project: Project = { id: 'ov-nosnap', name: '无快照', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:ovN', projectId: 'ov-nosnap', title: '无快照包' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-nosnap': { releasePackages: [release] } },
-      creatorPerformanceData: { 'ov-nosnap': { snapshots: [] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    const row = await waitFor(() => performanceOverviewRowByReleaseId(overview, 'creator-release:ovN'));
-    const tds = row.querySelectorAll('td');
-    expect(tds[3]!.textContent).toContain('No performance snapshots');
-    expect(tds[4]!.textContent?.trim()).toBe('-');
-    expect(tds[5]!.textContent?.trim()).toBe('-');
-    expect(tds[6]!.textContent?.trim()).toBe('-');
-  });
-
-  it('performance overview filters by platform across all five segments', async () => {
-    const project: Project = { id: 'ov-plat', name: '平台', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const rBili = makePublishedRelease({ id: 'creator-release:pBili', projectId: 'ov-plat', title: 'B站包', platform: 'bilibili' });
-    const rYt = makePublishedRelease({ id: 'creator-release:pYt', projectId: 'ov-plat', title: 'YT包', platform: 'youtube' });
-    const rXhs = makePublishedRelease({ id: 'creator-release:pXhs', projectId: 'ov-plat', title: '小红书包', platform: 'xiaohongshu' });
-    const rOther = makePublishedRelease({ id: 'creator-release:pOther', projectId: 'ov-plat', title: '其他包', platform: 'other' });
-    const snap = (releaseId: string, id: string) => makeSnapshot({ id, projectId: 'ov-plat', releaseId, capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 5 } });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-plat': { releasePackages: [rBili, rYt, rXhs, rOther] } },
-      creatorPerformanceData: { 'ov-plat': { snapshots: [snap('creator-release:pBili', 's1'), snap('creator-release:pYt', 's2'), snap('creator-release:pXhs', 's3'), snap('creator-release:pOther', 's4')] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(4));
-
-    fireEvent.click(within(overview).getByRole('button', { name: 'Bilibili' }));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:pBili']);
-    fireEvent.click(within(overview).getByRole('button', { name: 'YouTube' }));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:pYt']);
-    fireEvent.click(within(overview).getByRole('button', { name: 'Xiaohongshu' }));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:pXhs']);
-    fireEvent.click(within(overview).getByRole('button', { name: 'Other' }));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:pOther']);
-    fireEvent.click(within(overview).getByRole('button', { name: 'All' }));
-    expect(performanceOverviewReleaseOrder(overview).length).toBe(4);
-  });
-
-  it('performance overview sorts by latest snapshot, views, likes, and comments', async () => {
-    const project: Project = { id: 'ov-sort', name: '排序', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const r1 = makePublishedRelease({ id: 'creator-release:R1', projectId: 'ov-sort', title: '包一' });
-    const r2 = makePublishedRelease({ id: 'creator-release:R2', projectId: 'ov-sort', title: '包二' });
-    const r3 = makePublishedRelease({ id: 'creator-release:R3', projectId: 'ov-sort', title: '包三' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-sort': { releasePackages: [r1, r2, r3] } },
-      creatorPerformanceData: { 'ov-sort': { snapshots: [
-        makeSnapshot({ id: 'c1', projectId: 'ov-sort', releaseId: 'creator-release:R1', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 300, likes: 5, comments: 9 } }),
-        makeSnapshot({ id: 'c2', projectId: 'ov-sort', releaseId: 'creator-release:R2', capturedAt: '2026-07-05T00:00:00.000Z', metrics: { views: 100, likes: 50, comments: 1 } }),
-        makeSnapshot({ id: 'c3', projectId: 'ov-sort', releaseId: 'creator-release:R3', capturedAt: '2026-07-01T00:00:00.000Z', metrics: { views: 200, likes: 20, comments: 5 } }),
-      ] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(3));
-
-    // 默认按 Latest snapshot 降序（capturedAt）：R2(07-05) → R1(07-03) → R3(07-01)。
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:R2', 'creator-release:R1', 'creator-release:R3']);
-    const sortSelect = within(overview).getByLabelText('Performance overview sort');
-    fireEvent.change(sortSelect, { target: { value: 'views' } });
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:R1', 'creator-release:R3', 'creator-release:R2']);
-    fireEvent.change(sortSelect, { target: { value: 'likes' } });
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:R2', 'creator-release:R3', 'creator-release:R1']);
-    fireEvent.change(sortSelect, { target: { value: 'comments' } });
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:R1', 'creator-release:R3', 'creator-release:R2']);
-  });
-
-  it('performance overview orders releases missing the sort value after those with values', async () => {
-    const project: Project = { id: 'ov-miss-sort', name: '缺值排序', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const rWith = makePublishedRelease({ id: 'creator-release:with', projectId: 'ov-miss-sort', title: '有值包' });
-    const rNoView = makePublishedRelease({ id: 'creator-release:noview', projectId: 'ov-miss-sort', title: '无该指标包' });
-    const rNoSnap = makePublishedRelease({ id: 'creator-release:nosnap', projectId: 'ov-miss-sort', title: '无快照包' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-miss-sort': { releasePackages: [rWith, rNoView, rNoSnap] } },
-      creatorPerformanceData: { 'ov-miss-sort': { snapshots: [
-        makeSnapshot({ id: 'w1', projectId: 'ov-miss-sort', releaseId: 'creator-release:with', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 100 } }),
-        makeSnapshot({ id: 'n1', projectId: 'ov-miss-sort', releaseId: 'creator-release:noview', capturedAt: '2026-07-04T00:00:00.000Z', metrics: { likes: 9 } }),
-      ] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(3));
-
-    fireEvent.change(within(overview).getByLabelText('Performance overview sort'), { target: { value: 'views' } });
-    const order = performanceOverviewReleaseOrder(overview);
-    // 有 views 的排最前；缺失该指标（无该指标 / 无快照）的两条都排其后。
-    expect(order[0]).toBe('creator-release:with');
-    expect(order.slice(1)).toContain('creator-release:noview');
-    expect(order.slice(1)).toContain('creator-release:nosnap');
-  });
-
-  it('performance overview breaks sort ties deterministically by capturedAt, title, then id', async () => {
-    const project: Project = { id: 'ov-tie', name: '并列', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    // 四条 release 的 views 与 capturedAt 完全相同：先按标题升序（Alpha<Beta<Same），
-    // 同标题再按 id 升序（id-1<id-2）。乱序传入以证明排序确定性。
-    const rAlpha = makePublishedRelease({ id: 'creator-release:z-alpha', projectId: 'ov-tie', title: 'Alpha' });
-    const rBeta = makePublishedRelease({ id: 'creator-release:a-beta', projectId: 'ov-tie', title: 'Beta' });
-    const rSame1 = makePublishedRelease({ id: 'creator-release:id-1', projectId: 'ov-tie', title: 'Same' });
-    const rSame2 = makePublishedRelease({ id: 'creator-release:id-2', projectId: 'ov-tie', title: 'Same' });
-    const cap = '2026-07-02T00:00:00.000Z';
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-tie': { releasePackages: [rSame2, rBeta, rSame1, rAlpha] } },
-      creatorPerformanceData: { 'ov-tie': { snapshots: [
-        makeSnapshot({ id: 't-a', projectId: 'ov-tie', releaseId: 'creator-release:z-alpha', capturedAt: cap, metrics: { views: 100 } }),
-        makeSnapshot({ id: 't-b', projectId: 'ov-tie', releaseId: 'creator-release:a-beta', capturedAt: cap, metrics: { views: 100 } }),
-        makeSnapshot({ id: 't-c', projectId: 'ov-tie', releaseId: 'creator-release:id-1', capturedAt: cap, metrics: { views: 100 } }),
-        makeSnapshot({ id: 't-d', projectId: 'ov-tie', releaseId: 'creator-release:id-2', capturedAt: cap, metrics: { views: 100 } }),
-      ] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(4));
-    fireEvent.change(within(overview).getByLabelText('Performance overview sort'), { target: { value: 'views' } });
-    expect(performanceOverviewReleaseOrder(overview)).toEqual([
-      'creator-release:z-alpha',
-      'creator-release:a-beta',
-      'creator-release:id-1',
-      'creator-release:id-2',
-    ]);
-  });
-
-  it('performance overview excludes draft, ready, and archived releases', async () => {
-    const project: Project = { id: 'ov-status', name: '状态', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const draft = makeRelease({ id: 'creator-release:st-draft', projectId: 'ov-status', title: '草稿态', status: 'draft' });
-    const ready = makeRelease({ id: 'creator-release:st-ready', projectId: 'ov-status', title: '就绪态', status: 'ready' });
-    const archived = makeRelease({ id: 'creator-release:st-arch', projectId: 'ov-status', title: '归档态', status: 'archived' });
-    const published = makePublishedRelease({ id: 'creator-release:st-pub', projectId: 'ov-status', title: '发布态' });
-    const snap = (releaseId: string, id: string) => makeSnapshot({ id, projectId: 'ov-status', releaseId, capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 7 } });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-status': { releasePackages: [draft, ready, archived, published] } },
-      creatorPerformanceData: { 'ov-status': { snapshots: [snap('creator-release:st-draft', 'd'), snap('creator-release:st-ready', 'r'), snap('creator-release:st-arch', 'a'), snap('creator-release:st-pub', 'p')] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(1));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:st-pub']);
-    expect(within(overview).queryByText('草稿态')).toBeNull();
-    expect(within(overview).queryByText('就绪态')).toBeNull();
-    expect(within(overview).queryByText('归档态')).toBeNull();
-  });
-
-  it('performance overview drops a project when its release API fails', async () => {
-    const failing: Project = { id: 'ov-relfail', name: '发布失败项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const healthy: Project = { id: 'ov-relok', name: '健康项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const failRelease = makePublishedRelease({ id: 'creator-release:relfail', projectId: 'ov-relfail', title: '失败项目包' });
-    const okRelease = makePublishedRelease({ id: 'creator-release:relok', projectId: 'ov-relok', title: '健康项目包' });
-    mockTasksViewFetch({
-      creatorProjects: [failing, healthy],
-      creatorReleaseData: {
-        'ov-relfail': { releasePackages: [failRelease] },
-        'ov-relok': { releasePackages: [okRelease] },
-      },
-      creatorReleaseFailures: ['ov-relfail'],
-      creatorPerformanceData: {
-        'ov-relfail': { snapshots: [makeSnapshot({ id: 'rf', projectId: 'ov-relfail', releaseId: 'creator-release:relfail', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 1 } })] },
-        'ov-relok': { snapshots: [makeSnapshot({ id: 'ro', projectId: 'ov-relok', releaseId: 'creator-release:relok', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 2 } })] },
-      },
-    });
-
-    render(<TasksView projects={[failing, healthy]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(1));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:relok']);
-    expect(within(overview).queryByText('失败项目包')).toBeNull();
-    expect(within(overview).getByText('健康项目包')).toBeTruthy();
-  });
-
-  it('performance overview drops a project when its performance API fails', async () => {
-    const failing: Project = { id: 'ov-perffail', name: '表现失败项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const healthy: Project = { id: 'ov-perfok', name: '健康项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const failRelease = makePublishedRelease({ id: 'creator-release:perffail', projectId: 'ov-perffail', title: '表现失败包' });
-    const okRelease = makePublishedRelease({ id: 'creator-release:perfok', projectId: 'ov-perfok', title: '表现健康包' });
-    mockTasksViewFetch({
-      creatorProjects: [failing, healthy],
-      creatorReleaseData: {
-        'ov-perffail': { releasePackages: [failRelease] },
-        'ov-perfok': { releasePackages: [okRelease] },
-      },
-      creatorPerformanceData: {
-        'ov-perffail': { snapshots: [makeSnapshot({ id: 'pf', projectId: 'ov-perffail', releaseId: 'creator-release:perffail', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 1 } })] },
-        'ov-perfok': { snapshots: [makeSnapshot({ id: 'po', projectId: 'ov-perfok', releaseId: 'creator-release:perfok', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 2 } })] },
-      },
-      creatorPerformanceFailures: ['ov-perffail'],
-    });
-
-    render(<TasksView projects={[failing, healthy]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(1));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:perfok']);
-    expect(within(overview).queryByText('表现失败包')).toBeNull();
-    expect(within(overview).getByText('表现健康包')).toBeTruthy();
-  });
-
-  it('performance overview shows an unavailable hint when a project release API fails, keeps healthy projects, and leaves other panels usable', async () => {
-    const failing: Project = { id: 'ov-relfail2', name: '发布失败项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const healthy: Project = { id: 'ov-relok2', name: '健康项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const failRelease = makePublishedRelease({ id: 'creator-release:relfail2', projectId: 'ov-relfail2', title: '失败项目包' });
-    const okRelease = makePublishedRelease({ id: 'creator-release:relok2', projectId: 'ov-relok2', title: '健康项目包' });
-    mockTasksViewFetch({
-      creatorProjects: [failing, healthy],
-      creatorReleaseData: {
-        'ov-relfail2': { releasePackages: [failRelease] },
-        'ov-relok2': { releasePackages: [okRelease] },
-      },
-      creatorReleaseFailures: ['ov-relfail2'],
-      creatorPerformanceData: {
-        'ov-relfail2': { snapshots: [makeSnapshot({ id: 'rf2', projectId: 'ov-relfail2', releaseId: 'creator-release:relfail2', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 1 } })] },
-        'ov-relok2': { snapshots: [makeSnapshot({ id: 'ro2', projectId: 'ov-relok2', releaseId: 'creator-release:relok2', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 2 } })] },
-      },
-    });
-
-    render(<TasksView projects={[failing, healthy]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(1));
-    // 健康项目仍显示，失败项目不显示任何聚合数据。
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:relok2']);
-    expect(within(overview).queryByText('失败项目包')).toBeNull();
-    expect(within(overview).getByText('健康项目包')).toBeTruthy();
-    // 项目级降级提示可见，且为单数形式。
-    expect(within(overview).getByText('Performance overview unavailable for 1 project.')).toBeTruthy();
-    // 既有 Tasks 区域（Automations 子标签）仍可用，未被总览失败破坏。
-    expect(screen.getByRole('tab', { name: /Automations/i })).toBeTruthy();
-  });
-
-  it('performance overview shows an unavailable hint when a project performance API fails and keeps healthy projects', async () => {
-    const failing: Project = { id: 'ov-perffail2', name: '表现失败项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const healthy: Project = { id: 'ov-perfok2', name: '健康项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const failRelease = makePublishedRelease({ id: 'creator-release:perffail2', projectId: 'ov-perffail2', title: '表现失败包' });
-    const okRelease = makePublishedRelease({ id: 'creator-release:perfok2', projectId: 'ov-perfok2', title: '表现健康包' });
-    mockTasksViewFetch({
-      creatorProjects: [failing, healthy],
-      creatorReleaseData: {
-        'ov-perffail2': { releasePackages: [failRelease] },
-        'ov-perfok2': { releasePackages: [okRelease] },
-      },
-      creatorPerformanceData: {
-        'ov-perffail2': { snapshots: [makeSnapshot({ id: 'pf2', projectId: 'ov-perffail2', releaseId: 'creator-release:perffail2', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 1 } })] },
-        'ov-perfok2': { snapshots: [makeSnapshot({ id: 'po2', projectId: 'ov-perfok2', releaseId: 'creator-release:perfok2', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 2 } })] },
-      },
-      creatorPerformanceFailures: ['ov-perffail2'],
-    });
-
-    render(<TasksView projects={[failing, healthy]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(1));
-    expect(performanceOverviewReleaseOrder(overview)).toEqual(['creator-release:perfok2']);
-    expect(within(overview).queryByText('表现失败包')).toBeNull();
-    expect(within(overview).getByText('表现健康包')).toBeTruthy();
-    expect(within(overview).getByText('Performance overview unavailable for 1 project.')).toBeTruthy();
-  });
-
-  it('performance overview counts a project only once when both its release and performance APIs fail', async () => {
-    const failing: Project = { id: 'ov-bothfail', name: '双失败项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:bothfail', projectId: 'ov-bothfail', title: '双失败包' });
-    mockTasksViewFetch({
-      creatorProjects: [failing],
-      creatorReleaseData: { 'ov-bothfail': { releasePackages: [release] } },
-      creatorReleaseFailures: ['ov-bothfail'],
-      creatorPerformanceData: { 'ov-bothfail': { snapshots: [] } },
-      creatorPerformanceFailures: ['ov-bothfail'],
-    });
-
-    render(<TasksView projects={[failing] } />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    // 同一项目两个接口都失败 → 只计 1，不重复计数。
-    expect(await within(overview).findByText('Performance overview unavailable for 1 project.')).toBeTruthy();
-    expect(within(overview).queryByText('Performance overview unavailable for 2 projects.')).toBeNull();
-    expect(performanceOverviewReleaseOrder(overview).length).toBe(0);
-  });
-
-  it('performance overview keeps the unavailable hint and shows an available-releases empty state when every project fails', async () => {
-    const failing: Project = { id: 'ov-allfail', name: '全失败项目', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const release = makePublishedRelease({ id: 'creator-release:allfail', projectId: 'ov-allfail', title: '全失败包' });
-    mockTasksViewFetch({
-      creatorProjects: [failing],
-      creatorReleaseData: { 'ov-allfail': { releasePackages: [release] } },
-      creatorReleaseFailures: ['ov-allfail'],
-      creatorPerformanceData: { 'ov-allfail': { snapshots: [] } },
-      creatorPerformanceFailures: ['ov-allfail'],
-    });
-
-    render(<TasksView projects={[failing] } />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    // 失败提示必须保留。
-    expect(await within(overview).findByText('Performance overview unavailable for 1 project.')).toBeTruthy();
-    // 空态不得仅暗示“没有已发布 release”，必须表达为“没有可用已发布 release”。
-    expect(within(overview).getByText('No available published releases to compare yet.')).toBeTruthy();
-    expect(within(overview).queryByText('No published releases to compare yet.')).toBeNull();
-    expect(overview.querySelector('table')).toBeNull();
-  });
-
-  it('performance overview resolves content titles and falls back to the content id', async () => {
-    const project: Project = { id: 'ov-content', name: '内容标题', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const titled = makePublishedRelease({ id: 'creator-release:ct-titled', projectId: 'ov-content', title: '有标题包', contentId: 'creator-content:1' });
-    const fallback = makePublishedRelease({ id: 'creator-release:ct-fallback', projectId: 'ov-content', title: '回退包', contentId: 'creator-content:orphan' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-content': { releasePackages: [titled, fallback] } },
-      creatorContentData: { 'ov-content': { contentProjects: [{
-        id: 'creator-content:1', projectId: 'ov-content', title: '正片内容', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-      creatorPerformanceData: { 'ov-content': { snapshots: [
-        makeSnapshot({ id: 'ct1', projectId: 'ov-content', releaseId: 'creator-release:ct-titled', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 3 } }),
-      ] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(2));
-    const titledRow = performanceOverviewRowByReleaseId(overview, 'creator-release:ct-titled');
-    const fallbackRow = performanceOverviewRowByReleaseId(overview, 'creator-release:ct-fallback');
-    expect(titledRow.querySelectorAll('td')[1]!.textContent).toContain('正片内容');
-    // 内容无法解析时回退显示 contentId，且 release 仍然展示。
-    expect(fallbackRow.querySelectorAll('td')[1]!.textContent).toContain('creator-content:orphan');
-  });
-
-  it('performance overview shows an empty state when no published releases exist', async () => {
-    const project: Project = { id: 'ov-empty', name: '空态', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const draft = makeRelease({ id: 'creator-release:empty-draft', projectId: 'ov-empty', title: '仅草稿', status: 'draft' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-empty': { releasePackages: [draft] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    expect(await within(overview).findByText('No published releases to compare yet.')).toBeTruthy();
-    expect(overview.querySelector('table')).toBeNull();
-  });
-
-  it('performance overview interactions never issue write requests', async () => {
-    const project: Project = { id: 'ov-readonly', name: '只读', skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' } };
-    const rBili = makePublishedRelease({ id: 'creator-release:ro-bili', projectId: 'ov-readonly', title: '只读B站', platform: 'bilibili' });
-    const rYt = makePublishedRelease({ id: 'creator-release:ro-yt', projectId: 'ov-readonly', title: '只读YT', platform: 'youtube' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'ov-readonly': { releasePackages: [rBili, rYt] } },
-      creatorPerformanceData: { 'ov-readonly': { snapshots: [
-        makeSnapshot({ id: 'ro1', projectId: 'ov-readonly', releaseId: 'creator-release:ro-bili', capturedAt: '2026-07-02T00:00:00.000Z', metrics: { views: 4 } }),
-        makeSnapshot({ id: 'ro2', projectId: 'ov-readonly', releaseId: 'creator-release:ro-yt', capturedAt: '2026-07-03T00:00:00.000Z', metrics: { views: 6 } }),
-      ] } },
-    });
-
-    render(<TasksView projects={[project]} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const overview = await screen.findByRole('region', { name: 'Performance overview' });
-    await waitFor(() => expect(performanceOverviewReleaseOrder(overview).length).toBe(2));
-    expect(countWriteRequests()).toBe(0);
-
-    fireEvent.click(within(overview).getByRole('button', { name: 'Bilibili' }));
-    fireEvent.click(within(overview).getByRole('button', { name: 'YouTube' }));
-    fireEvent.click(within(overview).getByRole('button', { name: 'All' }));
-    fireEvent.change(within(overview).getByLabelText('Performance overview sort'), { target: { value: 'views' } });
-    fireEvent.change(within(overview).getByLabelText('Performance overview sort'), { target: { value: 'latest' } });
-
-    expect(countWriteRequests()).toBe(0);
-  });
-
-  it('performance overview ships responsive rules for the 960px and 640px breakpoints', () => {
-    const cssPath = join(dirname(fileURLToPath(import.meta.url)), '..', '..', 'src', 'styles', 'home', 'tasks.css');
-    const css = readFileSync(cssPath, 'utf8').replace(/\r\n/g, '\n');
-    expect(css).toContain('.creator-performance-overview__scroll {');
-    expect(css).toContain('.creator-performance-overview__table {');
-    expect(css).toContain('@media (max-width: 960px) {\n  .creator-performance-overview__scroll {');
-    expect(css).toContain('@media (max-width: 640px) {\n  .creator-performance-overview__head {');
-    expect(css).toContain('min-width: 640px;');
-  });
-
-  // ===== CW-06 Release schedule tests =====
-  const RS_NOW = new Date('2026-07-15T12:00:00.000Z');
-  const rsProject = (id: string, name: string): Project => ({
-    id, name, skillId: null, designSystemId: null, createdAt: Date.now(), updatedAt: Date.now(), metadata: { kind: 'video' },
-  });
-
-  it('release schedule lists only draft and ready releases that have a scheduledAt across multiple projects', async () => {
-    const projectA = rsProject('rs-proj-a', '项目甲');
-    const projectB = rsProject('rs-proj-b', '项目乙');
-    const draftA = makeRelease({ id: 'creator-release:rsDa', projectId: 'rs-proj-a', title: '甲草稿', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const readyB = makeRelease({ id: 'creator-release:rsRb', projectId: 'rs-proj-b', title: '乙就绪', status: 'ready', scheduledAt: '2026-07-22T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [projectA, projectB],
-      creatorReleaseData: {
-        'rs-proj-a': { releasePackages: [draftA] },
-        'rs-proj-b': { releasePackages: [readyB] },
-      },
-    });
-    render(<TasksView projects={[projectA, projectB]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(2));
-    expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:rsDa', 'creator-release:rsRb']);
-    expect(within(region).getByText('甲草稿')).toBeTruthy();
-    expect(within(region).getByText('乙就绪')).toBeTruthy();
-  });
-
-  it('release schedule excludes published, archived, and releases without a scheduledAt', async () => {
-    const project = rsProject('rs-proj-x', '排除项目');
-    const keep = makeRelease({ id: 'creator-release:rsKeep', projectId: 'rs-proj-x', title: '应保留', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const published = makeRelease({ id: 'creator-release:rsPub', projectId: 'rs-proj-x', title: '已发布排除', status: 'published', scheduledAt: '2026-07-10T09:00:00.000Z' });
-    const archived = makeRelease({ id: 'creator-release:rsArc', projectId: 'rs-proj-x', title: '已归档排除', status: 'archived', scheduledAt: '2026-07-11T09:00:00.000Z' });
-    const draftNoSched = makeRelease({ id: 'creator-release:rsNoSched', projectId: 'rs-proj-x', title: '草稿无时间', status: 'draft' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-x': { releasePackages: [keep, published, archived, draftNoSched] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:rsKeep']));
-    expect(within(region).queryByText('已发布排除')).toBeNull();
-    expect(within(region).queryByText('已归档排除')).toBeNull();
-    expect(within(region).queryByText('草稿无时间')).toBeNull();
-  });
-
-  it('release schedule excludes releases with an unparseable scheduledAt while keeping valid neighbors', async () => {
-    const project = rsProject('rs-proj-bad', '非法日期项目');
-    const keep = makeRelease({ id: 'creator-release:rsBadKeep', projectId: 'rs-proj-bad', title: '合法邻居', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const bad = makeRelease({ id: 'creator-release:rsBad', projectId: 'rs-proj-bad', title: '非法日期', status: 'draft', scheduledAt: 'not-a-date' });
-    const oob = makeRelease({ id: 'creator-release:rsBadOob', projectId: 'rs-proj-bad', title: '非法月份日期', status: 'ready', scheduledAt: '2026-13-45T10:00:00.000Z' });
-    const emptyStr = makeRelease({ id: 'creator-release:rsBadEmpty', projectId: 'rs-proj-bad', title: '空串日期', status: 'draft', scheduledAt: '' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-bad': { releasePackages: [keep, bad, oob, emptyStr] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:rsBadKeep']));
-    expect(within(region).queryByText('非法日期')).toBeNull();
-    expect(within(region).queryByText('非法月份日期')).toBeNull();
-    expect(within(region).queryByText('空串日期')).toBeNull();
-  });
-
-  it('release schedule orders items by scheduledAt ascending with a stable tie-break by title then id', async () => {
-    const project = rsProject('rs-proj-o', '排序项目');
-    const r1 = makeRelease({ id: 'creator-release:rs1', title: 'Alpha 07-18', status: 'draft', scheduledAt: '2026-07-18T00:00:00.000Z' });
-    const tieA = makeRelease({ id: 'creator-release:tieA', title: 'Bravo tie', status: 'ready', scheduledAt: '2026-07-19T00:00:00.000Z' });
-    const tieB = makeRelease({ id: 'creator-release:tieB', title: 'Charlie tie', status: 'ready', scheduledAt: '2026-07-19T00:00:00.000Z' });
-    const r2 = makeRelease({ id: 'creator-release:rs2', title: 'Delta 07-20', status: 'draft', scheduledAt: '2026-07-20T00:00:00.000Z' });
-    const r3 = makeRelease({ id: 'creator-release:rs3', title: 'Echo 07-22', status: 'draft', scheduledAt: '2026-07-22T00:00:00.000Z' });
-    // 乱序传入以证明不信任 API 原顺序。
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-o': { releasePackages: [r3, tieB, r1, tieA, r2] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(5));
-    // 升序：07-18 → 07-19(同时间按 title: Bravo < Charlie) → 07-20 → 07-22
-    expect(releaseScheduleItemOrder(region)).toEqual([
-      'creator-release:rs1',
-      'creator-release:tieA',
-      'creator-release:tieB',
-      'creator-release:rs2',
-      'creator-release:rs3',
-    ]);
-  });
-
-  it('release schedule groups items by local date', async () => {
-    const project = rsProject('rs-proj-g', '分组项目');
-    // 两个「同日」项使用接近正午的 UTC 时间，确保任何真实时区下都落在同一本地日（避免跨时区断言脆弱）。
-    const d1a = makeRelease({ id: 'creator-release:g1', title: '组一早', status: 'draft', scheduledAt: '2026-07-18T03:00:00.000Z' });
-    const d1b = makeRelease({ id: 'creator-release:g2', title: '组一晚', status: 'draft', scheduledAt: '2026-07-18T05:00:00.000Z' });
-    const d2 = makeRelease({ id: 'creator-release:g3', title: '组二', status: 'draft', scheduledAt: '2026-07-21T03:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-g': { releasePackages: [d2, d1b, d1a] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleGroupDates(region).length).toBe(2));
-    // 整体顺序保持 scheduledAt 升序，跨日期组连续。
-    expect(releaseScheduleItemOrder(region)).toEqual([
-      'creator-release:g1',
-      'creator-release:g2',
-      'creator-release:g3',
-    ]);
-    const groups = region.querySelectorAll('[data-testid="creator-release-schedule-group"]');
-    expect(within(groups[0] as HTMLElement).getByText('组一早')).toBeTruthy();
-    expect(within(groups[0] as HTMLElement).getByText('组一晚')).toBeTruthy();
-    expect(within(groups[1] as HTMLElement).getByText('组二')).toBeTruthy();
-  });
-
-  it('release schedule marks draft and ready releases scheduled in the past as Overdue', async () => {
-    const project = rsProject('rs-proj-od', '逾期项目');
-    const overdue = makeRelease({ id: 'creator-release:od', title: '逾期项', status: 'draft', scheduledAt: '2026-07-10T09:00:00.000Z' });
-    const future = makeRelease({ id: 'creator-release:fu', title: '未来项', status: 'ready', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-od': { releasePackages: [overdue, future] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(2));
-    expect(within(releaseScheduleItemByReleaseId(region, 'creator-release:od')).getByText('Overdue')).toBeTruthy();
-    expect(within(releaseScheduleItemByReleaseId(region, 'creator-release:fu')).queryByText('Overdue')).toBeNull();
-  });
-
-  it('release schedule Next 7 days filter includes now..now+7d and excludes the +7d boundary and farther items', async () => {
-    const project = rsProject('rs-proj-w7', '七天项目');
-    const within7 = makeRelease({ id: 'creator-release:w7', title: '七天内', status: 'draft', scheduledAt: '2026-07-18T12:00:00.000Z' });
-    const boundary7 = makeRelease({ id: 'creator-release:b7', title: '七天整', status: 'draft', scheduledAt: '2026-07-22T12:00:00.000Z' });
-    const far = makeRelease({ id: 'creator-release:far', title: '更远', status: 'draft', scheduledAt: '2026-07-25T12:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-w7': { releasePackages: [within7, boundary7, far] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(3));
-    fireEvent.click(within(region).getByRole('button', { name: 'Next 7 days' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:w7']));
-    expect(within(region).queryByText('七天整')).toBeNull();
-    expect(within(region).queryByText('更远')).toBeNull();
-  });
-
-  it('release schedule Next 30 days filter includes up to now+30d and excludes the +30d boundary', async () => {
-    const project = rsProject('rs-proj-w30', '三十天项目');
-    const within30 = makeRelease({ id: 'creator-release:w30', title: '三十天内', status: 'draft', scheduledAt: '2026-08-04T12:00:00.000Z' });
-    const boundary30 = makeRelease({ id: 'creator-release:b30', title: '三十天整', status: 'draft', scheduledAt: '2026-08-14T12:00:00.000Z' });
-    const past = makeRelease({ id: 'creator-release:od30', title: '过去', status: 'draft', scheduledAt: '2026-07-10T12:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-w30': { releasePackages: [within30, boundary30, past] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(3));
-    fireEvent.click(within(region).getByRole('button', { name: 'Next 30 days' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:w30']));
-    expect(within(region).queryByText('三十天整')).toBeNull();
-    expect(within(region).queryByText('过去')).toBeNull();
-  });
-
-  it('release schedule Overdue filter shows only releases scheduled before now', async () => {
-    const project = rsProject('rs-proj-of', '逾期筛选项目');
-    const past = makeRelease({ id: 'creator-release:op', title: '过去项', status: 'draft', scheduledAt: '2026-07-10T12:00:00.000Z' });
-    const future = makeRelease({ id: 'creator-release:of', title: '未来项', status: 'draft', scheduledAt: '2026-07-20T12:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-of': { releasePackages: [past, future] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(2));
-    fireEvent.click(within(region).getByRole('button', { name: 'Overdue' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:op']));
-    expect(within(region).queryByText('未来项')).toBeNull();
-  });
-
-  it('release schedule platform filter shows only the selected platform across all five segments', async () => {
-    const project = rsProject('rs-proj-p', '平台项目');
-    const bili = makeRelease({ id: 'creator-release:pB', title: 'B站', platform: 'bilibili', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const yt = makeRelease({ id: 'creator-release:pY', title: '油管', platform: 'youtube', status: 'draft', scheduledAt: '2026-07-20T10:00:00.000Z' });
-    const xhs = makeRelease({ id: 'creator-release:pX', title: '小红书', platform: 'xiaohongshu', status: 'draft', scheduledAt: '2026-07-20T11:00:00.000Z' });
-    const other = makeRelease({ id: 'creator-release:pO', title: '其他', platform: 'other', status: 'draft', scheduledAt: '2026-07-20T12:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-p': { releasePackages: [bili, yt, xhs, other] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(4));
-    fireEvent.click(within(region).getByRole('button', { name: 'YouTube' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:pY']));
-    fireEvent.click(within(region).getByRole('button', { name: 'Xiaohongshu' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:pX']));
-    fireEvent.click(within(region).getByRole('button', { name: 'Other' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:pO']));
-    fireEvent.click(within(region).getByRole('button', { name: 'Bilibili' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:pB']));
-    fireEvent.click(within(region).getByRole('button', { name: 'All' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(4));
-  });
-
-  it('release schedule resolves content titles and falls back to the content id', async () => {
-    const project = rsProject('rs-proj-c', '内容项目');
-    const titled = makeRelease({ id: 'creator-release:cw', title: '有内容', contentId: 'creator-content:alpha', platform: 'bilibili', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const fallback = makeRelease({ id: 'creator-release:cn', title: '无内容', contentId: 'creator-content:orphan', platform: 'youtube', status: 'ready', scheduledAt: '2026-07-20T10:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-c': { releasePackages: [titled, fallback] } },
-      creatorContentData: { 'rs-proj-c': { contentProjects: [{
-        id: 'creator-content:alpha', projectId: 'rs-proj-c', title: 'Alpha 内容', status: 'drafting', brief: {}, outline: {}, retrospective: {}, taskIds: [], storyboardItems: [],
-        createdAt: '2025-01-01T00:00:00Z', updatedAt: '2025-01-01T00:00:00Z',
-      }] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(2));
-    expect(within(region).getByText('Alpha 内容')).toBeTruthy();
-    expect(within(region).getByText('creator-content:orphan')).toBeTruthy();
-  });
-
-  it('release schedule shows an empty state when active filters match no releases', async () => {
-    const project = rsProject('rs-proj-e', '空态项目');
-    const far = makeRelease({ id: 'creator-release:e1', title: '远期', status: 'draft', scheduledAt: '2026-08-01T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-e': { releasePackages: [far] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(1));
-    fireEvent.click(within(region).getByRole('button', { name: 'Next 7 days' }));
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(0));
-    expect(within(region).getByText('No scheduled releases to show yet.')).toBeTruthy();
-  });
-
-  it('release schedule drops a project when its release API fails, keeps healthy projects, shows an unavailable hint, and leaves the Tasks panel usable', async () => {
-    const healthy = rsProject('rs-healthy', '健康项目');
-    const failing = rsProject('rs-fail', '失败项目');
-    const hRelease = makeRelease({ id: 'creator-release:rh', projectId: 'rs-healthy', title: '健康排期', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const fRelease = makeRelease({ id: 'creator-release:rf', projectId: 'rs-fail', title: '失败排期', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [healthy, failing],
-      creatorReleaseData: {
-        'rs-healthy': { releasePackages: [hRelease] },
-        'rs-fail': { releasePackages: [fRelease] },
-      },
-      creatorReleaseFailures: ['rs-fail'],
-    });
-    render(<TasksView projects={[healthy, failing]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region)).toEqual(['creator-release:rh']));
-    expect(within(region).getByText('Release schedule unavailable for 1 project.')).toBeTruthy();
-    expect(within(region).queryByText('失败排期')).toBeNull();
-    // 既有 Tasks 区域仍可用：Automations 标签仍在。
-    expect(screen.getByRole('tab', { name: /Automations/i })).toBeTruthy();
-  });
-
-  it('release schedule shows a plural hint for multiple failed projects', async () => {
-    const p1 = rsProject('rs-f1', '失败一');
-    const p2 = rsProject('rs-f2', '失败二');
-    const r1 = makeRelease({ id: 'creator-release:f1', projectId: 'rs-f1', title: '排期一', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const r2 = makeRelease({ id: 'creator-release:f2', projectId: 'rs-f2', title: '排期二', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [p1, p2],
-      creatorReleaseData: { 'rs-f1': { releasePackages: [r1] }, 'rs-f2': { releasePackages: [r2] } },
-      creatorReleaseFailures: ['rs-f1', 'rs-f2'],
-    });
-    render(<TasksView projects={[p1, p2]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(within(region).getByText('Release schedule unavailable for 2 projects.')).toBeTruthy());
-    expect(releaseScheduleItemOrder(region).length).toBe(0);
-  });
-
-  it('release schedule counts a project only once when both its release and performance APIs fail', async () => {
-    const project = rsProject('rs-both', '双失败项目');
-    const release = makeRelease({ id: 'creator-release:both', projectId: 'rs-both', title: '双失败排期', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-both': { releasePackages: [release] } },
-      creatorReleaseFailures: ['rs-both'],
-      creatorPerformanceFailures: ['rs-both'],
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    // 同一项目两个接口都失败 → 只计一次（1 project.，不是 2 projects.）。
-    await waitFor(() => expect(within(region).getByText('Release schedule unavailable for 1 project.')).toBeTruthy());
-    expect(releaseScheduleItemOrder(region).length).toBe(0);
-  });
-
-  it('release schedule keeps the unavailable hint and shows an available-releases empty state when every project fails', async () => {
-    const project = rsProject('rs-allfail', '全失败项目');
-    const release = makeRelease({ id: 'creator-release:af', projectId: 'rs-allfail', title: '全失败排期', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-allfail': { releasePackages: [release] } },
-      creatorReleaseFailures: ['rs-allfail'],
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(within(region).getByText('Release schedule unavailable for 1 project.')).toBeTruthy());
-    // 所有项目失败 → 保留提示，且空态不得仅暗示“没有排期”，必须表达为“没有可用排期”。
-    expect(within(region).getByText('No available scheduled releases to show.')).toBeTruthy();
-    expect(within(region).queryByText('No scheduled releases to show yet.')).toBeNull();
-    expect(region.querySelector('.creator-release-schedule__agenda')).toBeNull();
-  });
-
-  it('release schedule interactions never issue write requests', async () => {
-    const project = rsProject('rs-proj-z', '零写项目');
-    const bili = makeRelease({ id: 'creator-release:zB', title: 'B站', platform: 'bilibili', status: 'draft', scheduledAt: '2026-07-20T09:00:00.000Z' });
-    const yt = makeRelease({ id: 'creator-release:zY', title: '油管', platform: 'youtube', status: 'draft', scheduledAt: '2026-07-25T09:00:00.000Z' });
-    mockTasksViewFetch({
-      creatorProjects: [project],
-      creatorReleaseData: { 'rs-proj-z': { releasePackages: [bili, yt] } },
-    });
-    render(<TasksView projects={[project]} now={RS_NOW} />);
-    fireEvent.click(await screen.findByRole('tab', { name: /Creator workbench/i }));
-    const region = await screen.findByRole('region', { name: 'Release schedule' });
-    await waitFor(() => expect(releaseScheduleItemOrder(region).length).toBe(2));
-    fireEvent.click(within(region).getByRole('button', { name: 'YouTube' }));
-    fireEvent.click(within(region).getByRole('button', { name: 'Next 7 days' }));
-    fireEvent.click(within(region).getByRole('button', { name: 'Overdue' }));
-    fireEvent.click(within(region).getByRole('button', { name: 'All' }));
-    expect(countWriteRequests()).toBe(0);
-  });
-
-  it('release schedule ships responsive rules for the 960px and 640px breakpoints', () => {
-    const cssPath = join(dirname(fileURLToPath(import.meta.url)), '..', '..', 'src', 'styles', 'home', 'tasks.css');
-    const css = readFileSync(cssPath, 'utf8').replace(/\r\n/g, '\n');
-    expect(css).toContain('.creator-release-schedule {');
-    expect(css).toContain('.creator-release-schedule__agenda {');
-    expect(css).toContain('@media (max-width: 960px) {\n  .creator-release-schedule__head {');
-    expect(css).toContain('@media (max-width: 640px) {\n  .creator-release-schedule__item {');
-  });
 });
